Update go.mod to use go 1.21 and toolchain 1.22.3
This commit updates the `go.mod` file to set Go 1.21 as the minimum version
requirement for compatibility with dependent modules, and specifies Go 1.22.3
as the preferred toolchain to ensure that this version is used when building
binaries.

Setting the minimum version to go 1.21 instead of being explicit (e.g.: 1.21.3)
allows dependents to upgrade at their own pace. This approach ensures they can
use any patch release within the 1.21 series.